Svg (#156)
* Add SVG Test page

This adds support for react-native-svg on the js side. Win32 support for react-native-svg primitives are already checked into netui so this will light up support here.

Metro config files only allow for a single transformer to be specified so I had to create a transformer called transform-selector to select between react-native-transformer and the regular metro transformer depending on file extension. React-native-transformer transforms the xml from a svg file to javascript react-native calls at bundle-time so it doesn't have to be done at run-time.

A test page has been added to RNTester with examples for inline svgs, remotely retrieved svgs, and bundled svgs.

* Add SVG test missed files

Missed a few files on last commit.

* Addressed pr comments

* yarn.lock

* Change files