Create test-tools project with mock theme and some component tests (#495)
### Platforms Impacted
- [ ] iOS
- [ ] macOS
- [ ] win32 (Office)
- [ ] windows
- [ ] android

### Description of changes

This change adds a new package under utils called test-tools. This is a non-published, devDependency only, package for component testing. Right now it only contains a few things:

#### mock theme

This is just a fully specified theme that is exported and can be pulled into component tests if needed. There were a couple of locations where the theme was specified in tests so they were changed to point at the mock theme. This also allows some portions of the tests to be changed in one location.

#### enzyme component tests

This change also creates two related component tests that components we build can use to validate their correctness. They are `checkRenderConsistency` and `checkReRender`. The differences and similarities are as follows:

Scenario | `checkRenderConsistency` | `checkReRender`
--- | --- | ---
Scenario | Ensure that two identical instances of a component share as much as possible | Ensure that re-render doesn't cause unnecessary tree mutations 
Description | Calls the passed in render function twice and compares the results | Calls the render function, then triggers a tree update and re-render, and compares the results
Props | Props at each level are shallow compared using === | Props at each level are shallow compared using ===
Children | Ignored in the comparison | Ignored in the comparison
Functions | Ignored. Typically these might be built via useEffect hooks and will be unique per instance | Compared. Re-render should not rebuild functions passed to child components
Depth | 1 by default, set to 2 or more for HOCs. | 1 by default, set to 2 or more for HOCs

### Verification

(how the change was tested, including both manual and automated tests)

| Before                                       | After                                      |
|----------------------------------------------|--------------------------------------------|
| Screenshot or description before this change | Screenshot or description with this change |

Just tests, since this is a change to add tests

### Pull request checklist

This PR has considered (when applicable):
- [x] Automated Tests
- [ ] Documentation and examples
- [ ] Keyboard Accessibility
- [ ] Voiceover
- [ ] Internationalization and Right-to-left Layouts