Fixing redbox on macOS from theming module (#510)
* Fixing redbox on macOS from theming module

macOS doesn't currently provide a theme (the theming support on macOS is limited).

I think with commit 59d0857d4244fb6b35fb2c9ca7d01cf2a073f461 we can now end up using the spread operator on a string to contruct an object, which seems unintentional, since the following code

`Object.assign(translateOfficeTheme(themingModule, paletteCache, themeId), { name: _hostTheme })`

changed to

`{ /*other stuff*/, ...(_hostTheme && { name: _hostTheme }) }`

which is problematic because the empty string '' is falsy -- so instead of skipping the name field, we just end up trying to spread the empty string and trying to construct an object out of it, which the Object.es6.js polyfill is very unhappy about. This isn't a problem on Win32 because the theme is not falsy (i.e. not empty) on Win32.